+++ domains/certificates/Query.API/API/Query.API/ApiModels/CertificateList.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using System.Linq;
+using API.Query.API.ApiModels;
 
 namespace API.Models;
