+++ domains/certificates/Query.API/API/Query.API/Controllers/CertificatesController.cs
@@ -6,7 +6,7 @@ using Marten;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
-namespace API.Controllers;
+namespace API.Query.API.Controllers;
 
 [Authorize]
 [ApiController]
